requirements.txt
flask
gunicorn
gevent
tensorflow
Pillow
cassandra-driver==3.11.0